Refactor the pool2d_fwd example and add example for float type testing
